Non-working commented code was removed
An external javascript page was added to assign a
username and if this had been left blank it would
be assigned to stranger. This was really just to
play with code but a function to better decide if
know who the user is will be added later.

In the About Me page, two javascript functions
were added to update the contents of heading one
and two. The updates use the variable username
from the external javascript page. These take place
after the document finishes loading.